refactor(clinics): move persistence to infrastructure
