+++ server/features/clinics/infrastructure/index.ts
@@ -1,0 +1,1 @@
+export * from "./admin-clinics-repository.ts";
